# doc/apifox/openapi.yaml
openapi: 3.0.3
info:
  title: 高标农田专项整治后台 API
  description: |
    Go 1.27 · Gin + Zap + MySQL + JWT。
    统一响应：`{ code, data, message, cost_ms, trace_id }`，成功 `code=0`。
  version: 1.0.0
servers:
  - url: http://127.0.0.1:8080
    description: 本地开发
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPath:
      in: path
      name: id
      required: true
      schema: { type: integer, format: int64 }
...
  schemas:
    Envelope:
      type: object
      properties:
        code: { type: integer, example: 0 }
        data: {}
        message: { type: string, example: ok }
        cost_ms: { type: integer, example: 12 }
        trace_id: { type: string }
    LoginReq:
      type: object
      required: [username, password]
      properties:
        username: { type: string, example: admin }
        password: { type: string, example: "123456" }
    IssueInput:
      type: object
      properties:
        type: { type: string, enum: [well, road, bridge, forest, transformer] }
        street: { type: string }
        village: { type: string }
        project_name: { type: string }
        code: { type: string }
        location_text: { type: string }
        address: { type: string }
        lat: { type: number }
        lng: { type: number }
        description: { type: string }
        measures: { type: string }
        plan_date: { type: string, example: "2026-09-01" }
        reporter_name: { type: string }
        reporter_phone: { type: string }
        assignee_name: { type: string }
        assignee_phone: { type: string }
        file_uuids:
          type: array
          description: 文件 uuid 列表（新建必填；修改且 photo_ref_uuid 为空时必填）
          items: { type: string, format: uuid }
        photo_ref_uuid:
          type: string
          description: 修改时有值=附件不变；为空则用 file_uuids 重新关联
        type_ext: { type: object }
...
security:
  - bearerAuth: []
paths:
...
  /api/auth/login:
    post:
      tags: [鉴权]
      summary: 登录
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginReq" }
      responses:
        "200":
          description: 返回 token 与用户信息
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Envelope" }
...
  /api/issues:
    get:
      tags: [专项整改]
      summary: 问题列表
      parameters:
        - in: query
          name: type
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string, enum: [pending, done] }
        - in: query
          name: street
          schema: { type: string }
        - in: query
          name: village
          schema: { type: string }
        - in: query
          name: keyword
          schema: { type: string }
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: size
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Envelope" }
    post:
      tags: [专项整改]
      summary: 新增问题（提交即待整改）
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/IssueInput" }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Envelope" }
  /api/issues/{id}:
    get:
      tags: [专项整改]
      summary: 问题详情
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Envelope" }
    put:
      tags: [专项整改]
      summary: 更新问题
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/IssueInput" }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Envelope" }
    delete:
      tags: [专项整改]
      summary: 删除问题
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Envelope" }
  /api/issues/{id}/rectify:
    post:
      tags: [专项整改]
      summary: 提交整改结果
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                note: { type: string }
                file_uuids:
                  type: array
                  items: { type: string }
                rectify_photo_ref_uuid:
                  type: string
                  description: 有值=不变；空则用 file_uuids 重新关联
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Envelope" }
...
  /api/app/todos:
    get:
      tags: [小程序]
      summary: 待办列表
      description: |
        未传 status 时默认 pending；传 status=all 表示不限状态。
        筛选对齐 miniapp/todo：type / street / village / keyword / page / size。
      parameters:
        - in: query
          name: type
          schema: { type: string }
        - in: query
          name: status
          description: pending|done|all；省略则 pending
          schema: { type: string }
        - in: query
          name: street
          schema: { type: string }
        - in: query
          name: village
          schema: { type: string }
        - in: query
          name: keyword
          schema: { type: string }
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: size
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: data.list / total / page / size / status
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Envelope" }
...
  /api/app/issues:
    post:
      tags: [小程序]
      summary: 上报问题
      description: 提交即待整改；file_uuids → photo_ref_uuid。附件先调 /api/attachments/*。
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/IssueInput" }
      responses:
        "200":
          description: 新建问题（含 photos）
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Envelope" }
...